Tear out 32-bit APIs and frameworks. Replace RBSplitView with NSSplitView. Remove Sparkle.

+++ Other Sources/Foreign Code/SBCenteringClipView/SBCenteringClipView.m
@@ -90,7 +90,7 @@
     [self centerDocument];
 }
 
-- (void)setFrameRotation:(float)angle
+- (void)setFrameRotation:(CGFloat)angle
 {
     [super setFrameRotation:angle];
     [self centerDocument];
