+++ Pixel Art Core/Canvas/PXCanvas_Drawing.h
@@ -12,7 +12,7 @@
 @interface PXCanvas(Drawing)
 
 - (void)drawRect:(NSRect)rect;
-- (void)drawInRect:(NSRect)dst fromRect:(NSRect)src operation:(NSCompositingOperation)op fraction:(float)frac;
+- (void)drawInRect:(NSRect)dst fromRect:(NSRect)src operation:(NSCompositingOperation)op fraction:(CGFloat)frac;
 - (void)drawInRect:(NSRect)dst fromRect:(NSRect)src operation:(NSCompositingOperation)op;
 - (void)drawInRect:(NSRect)dst fromRect:(NSRect)src;
 - (void)meldBezier:(NSBezierPath *)path ofColor:(NSColor *)color;
